requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
spacy==3.7.6
Pillow==10.4.0
lxml==5.3.0
jinja2==3.1.4
google-cloud-aiplatform==1.71.1
